requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
python-socketio>=5.10.0
werkzeug>=3.0.0
watchdog>=4.0.0
gevent>=24.0.0
gevent-websocket>=0.10.1
gunicorn>=21.0.0
requests
apscheduler>=3.10.0